Application is possibly complete!
I have removed the Search functions, as they're not really that useful - and if you need them that badly, you can use the sample application from Front.

Cleaned up some of the extra files that came with different IDEs I had used.

Added a progress bar that updates as the export is performed.

Added a logo, and updated the README so it's much more useful