Better Handle Exports
Major update to this code, making it a more fully fledged application.

It is now able to handle partial Exports, and resume progress.

This is very handy when doing a large export and there is an error (for example, your internet goes down).

You don't have to start over, you can resume from where you got to!!